Move the api trace methods into a separate controller under the api namespace
